Add ShowActivityRecordsCommand and activity-records.jsp drafts

+++ src/main/java/org/oddys/timetracking/dao/ActivityRecordDao.java
@@ -6,7 +6,7 @@ import org.oddys.timetracking.entity.ActivityRecord;
 import java.util.List;
 
 public interface ActivityRecordDao {
-    int getNumberOfRows() throws DaoException;
+    long getNumberOfRows() throws DaoException;
 
     List<ActivityRecord> findActivityRecords(int currentPage, int RecordsPerPage) throws DaoException;
 }
