+++ src/main/java/org/oddys/timetracking/service/ActivityRecordService.java
@@ -1,5 +1,5 @@
 package org.oddys.timetracking.service;
 
 public interface ActivityRecordService {
-    int getNumberOfRows() throws ServiceException;
+    long getNumberOfRows() throws ServiceException;
 }
